feat: Introducing a React router!
