+++ package.json
@@ -31,6 +31,7 @@
     "pyodide": "^0.26.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "vite-plugin-pwa": "^0.20.5"
   },
   "devDependencies": {
