+++ src/constants.ts
@@ -1,0 +1,6 @@
+export const localStorageKeys = {
+  code: {
+    javascriptLastEdited: 'code:javascript:last-edited',
+    pythonLastEdited: 'code:python:last-edited',
+  },
+};
