+++ src/pages/Err404Page.tsx
@@ -1,0 +1,3 @@
+export default function Err404Page() {
+  return <p>404 not found</p>;
+}
